modification

+++ 12b.pointer_functions.c
@@ -1,13 +1,13 @@
 #include<stdio.h>
 
 int mathadd(int num1,int num2);
-int mathavg(int a, int b);
+float mathavg(int a, int b);
 a=7;b=10;
 
 int main()
 {
     printf("The addition of two numbers is %d \n",mathadd(5,6));
-    printf("The average of two numbers is %d \n",mathavg(a,b));
+    printf("The average of two numbers is %f \n",mathavg(a,b));
     return 0;
 }
 
@@ -17,7 +17,7 @@ int mathadd(int num1,int num2)
     return result;
 }
 
-int mathavg(int a,int b)
+float mathavg(int a,int b)
 {
     return (a/b);
 }
